requirements.txt
Flask
flask-socketio
Flask-SQLAlchemy
mysql.connector
flask-mysqldb
Flask-ULID
aioboto3
aiobotocore
aiofiles==23.2.1
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
attrs==23.2.0
boto3
botocore
colorama==0.4.6
frozenlist==1.4.1
idna==3.7
jmespath==1.0.1
multidict==6.0.5
python-dateutil==2.9.0.post0
s3transfer==0.10.1
six==1.16.0
tqdm==4.66.4
urllib3==2.2.2
wrapt==1.16.0
yarl==1.9.4
